% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{add_watershed_hex_overlay}
\alias{add_watershed_hex_overlay}
\title{Add Hexagonal Overlay to Watershed}
\usage{
add_watershed_hex_overlay(huc_info, hex_diameter = 0.01)
}
\arguments{
\item{huc_info}{A watershed list returned from `get_watershed()`.}

\item{hex_diameter}{Numeric representing the diameter of the hexagons in CRS units.}
}
\value{
An S3 object of class `watershed_hex_overlay` containing the original geometry plus the hex layer.
}
\description{
Generates a spatial hexagonal grid across the target watershed bounds.
}
