% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrate_triangle.R
\name{autoplot.substrate}
\alias{autoplot.substrate}
\title{Plot Visual Substrate Network Layouts}
\usage{
\method{autoplot}{substrate}(object, ...)
}
\arguments{
\item{object}{Data structure of type `substrate` mapped previously via abstract tridiagonal coordinates.}

\item{...}{Additional graphical tracking arguments passing to generic rendering matrices.}
}
\value{
A `ggplot` visual abstraction of the Ewing network.
}
\description{
Organizes numerical nodes logically over explicit geometrical borders defined by `create_substrate` attributes.
}
