% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{autoplot.watershed_hex_overlay}
\alias{autoplot.watershed_hex_overlay}
\title{Plot Watershed Hexagonal Overlay}
\usage{
\method{autoplot}{watershed_hex_overlay}(object, ...)
}
\arguments{
\item{object}{An S3 object of class `watershed_hex_overlay`.}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
A `ggplot` object representing the spatial mesh.
}
\description{
Visualizes the watershed bounded area and the superimposed hexagonal scaling network.
}
