% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{build_base_map}
\alias{build_base_map}
\title{Build Interactive Base Map}
\usage{
build_base_map()
}
\value{
A `leaflet` HTML widget object
}
\description{
Constructs a blank interactive Leaflet mapping canvas optimized for North America
geographic search. It embeds the `leaflet.extras::addSearchOSM()` search bar,
allowing text-based flying capabilities (Option A interaction paradigm).
}
