% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrate_triangle.R
\name{create_substrate}
\alias{create_substrate}
\title{Create Substrate Geometric Object}
\usage{
create_substrate(topology, width = 10, step = 1)
}
\arguments{
\item{topology}{Abstract parameter list derived from `substrate_topology()`.}

\item{width}{Physical geometric limits of bounding constraints.}

\item{step}{Inner plotting density logic parameter.}
}
\value{
An S3 object of format class `substrate`, storing resolved attributes.
}
\description{
Iterates against a defined configuration metric to build explicit spatial network definitions
tracking Euclidean point plots, dynamic border labels, and spanning geometric boundary polygons.
}
