% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadInput}
\title{Download File App}
\usage{
downloadApp(title = "Download Options")

downloadServer(
  id,
  sim_data,
  distplot = shiny::reactive(NULL),
  sppplot = shiny::reactive(NULL),
  envplot = shiny::reactive(NULL)
)

downloadInput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{sim_data}{reactive list holding the primary `simApp` state values (`simres`, `nsim`)}

\item{distplot}{reactive evaluating to the main age class visualization}

\item{sppplot}{reactive evaluating to the host/parasite substrate plot grids}

\item{envplot}{reactive evaluating to the generalized statistical envelopes}
}
\description{
Shiny application components for exporting evaluation metrics, data tables, and dynamic plot
graphics out of the simulation runtime.
}
