% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envPlotApp.R
\name{envPlotApp}
\alias{envPlotApp}
\alias{envPlotServer}
\alias{envPlotInput}
\alias{envPlotOutput}
\title{Envelope Plot App}
\usage{
envPlotApp(title = "Envelope Plots")

envPlotServer(id, simres, nsim)

envPlotInput(id)

envPlotOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{simres}{reactive holding evaluated multi-simulation responses}

\item{nsim}{reactive integer specifying iterations computed within the `simres` object}
}
\description{
Shiny application modules used strictly to render and wrap statistical envelope probability
intervals generated natively by active `nsim > 1` discrete operations.
}
