\name{ewing-internal}
\alias{activeTemp}
\alias{ageclass.plot}
\alias{break.backSpline}
\alias{car2tri.default}
\alias{car2tri}
\alias{cardist}
\alias{cdf.lines}
\alias{checkTime}
\alias{copyOrgInfo}
\alias{count.join}
\alias{curve.plot}
\alias{event.feed}
\alias{event.find}
\alias{event.move}
\alias{event.ovip}
\alias{fini.timing}
\alias{five.find}
\alias{five.lines}
\alias{five.make}
\alias{five.oplot}
\alias{five.plot}
\alias{five.show}
\alias{five.switch}
\alias{gasket}
\alias{get.alive}
\alias{get.attack}
\alias{get.base}
\alias{get.birth}
\alias{get.deplete}
\alias{get.future}
\alias{get.individual}
\alias{get.interact}
\alias{get.offspring}
\alias{get.species}
\alias{get.species.element}
\alias{getCount}
\alias{getDegreeDay}
\alias{getHour}
\alias{getOffspring}
\alias{getOrgAgeClass}
\alias{getOrgAlive}
\alias{getOrgFeature}
\alias{getOrgFuture}
\alias{getOrgHosts}
\alias{getOrgInfo}
\alias{getOrgInteract}
\alias{getOrgMeanValue}
\alias{getOrgSubstrate}
\alias{getTemp}
\alias{getTime}
\alias{host.ecto}
\alias{host.endo}
\alias{init.timing}
\alias{initCount}
\alias{initOffspring}
\alias{initOrgInfo}
\alias{initTemp}
\alias{is.move}
\alias{is.mydata}
\alias{leftist.birth}
\alias{leftist.create}
\alias{leftist.free}
\alias{leftist.merge}
\alias{leftist.remove}
\alias{leftist.tree}
\alias{leftist.update}
\alias{logpretty}
\alias{my.eval}
\alias{mydata}
\alias{plot.current}
\alias{parent.birth}
\alias{put.base}
\alias{put.individual}
\alias{put.species}
\alias{ramp.backSpline}
\alias{readCount}
\alias{rescale.temp}
\alias{ring.add}
\alias{ring.remove}
\alias{rspline}
\alias{rtri}
\alias{sampleOrgSubstrate}
\alias{set.birth}
\alias{set.future}
\alias{set.male}
\alias{set.offspring}
\alias{set.step}
\alias{set.timing}
\alias{setCount}
\alias{setEvents}
\alias{setOrgInfo}
\alias{setOrgMeanValue}
\alias{setTemp}
\alias{showTemp}
\alias{sierpinski}
\alias{spline.deriv}
\alias{spline.design}
\alias{spline.extrapolate}
\alias{spline.meanvalue}
\alias{spline.rate}
\alias{splinesum}
\alias{substrate.plot}
\alias{summaryshow}
\alias{temp.lines}
\alias{temp.repeat}
\alias{temp.spline}
\alias{text.current}
\alias{transTime}
\alias{tri2car.default}
\alias{tri2car}
\alias{tridist}
\alias{update.mintime}
\alias{updateCounts}
\alias{updateCount}
\alias{updateEvents}
\alias{updateTemp}
\alias{writeCount}
\title{Internal ewing functions.}
\description{  These are generally not to be called by the user.
}
\usage{
activeTemp(community,lo.hour,hi.hour,degreeday)
ageclass.plot(community,count,mfcol,main,col,cex,debugit)
break.backSpline(tmp)
car2tri(xy,xmult,ymult,sq3)
car2tri.default(x,y)
cardist(xy)
cdf.lines(data,fig,nspline,conf,rescale,col)
checkTime(community,x,base,units)
copyOrgInfo(fromname,toname)
count.join(...)
curve.plot(xy,y,z,n,action,fit,backfit,save.ends,col,lwd,f,finv)
event.feed(community,species,host,dead)
event.find(community,species,host,event)
event.ovip(community,species,host,dead,gender)
fini.timing(community)
five.find(fit,pick,vals,goal,refmean,tol,printit)
five.lines(invmvalue,dispersion,location,intensity,truncation,rejection,u,lty)
five.make(fit,dispersion,location,intensity,truncation,rejection,fivenum,u)
five.oplot(invmvalue,dispersion,location,intensity,truncation,rejection,ylim,u)
five.plot(gencurve,fit,pick,vals,ylim)
five.show(fit,goal,tol,legend.flag,cex,ylim,prefix)
five.switch(fit,pick,vals)
gasket(aa,bb)
get.base(community,species)
get.birth(community,species,offspring)
get.future(community,species,individuals,id)
getCount(community,species,element)
getDegreeDay(community,hour)
getHour(community,dd)
getOffspring(community,species,offspring)
getOrgAgeClass(community,species,stage,future)
getOrgAlive(community,species,element)
getOrgFeature(community,species,feature)
getOrgFuture(community,species,feature,current,future)
getOrgHosts(community,species,feature)
getOrgInfo(community,element)
getOrgInteract(community,org1name,org2name,event)
getOrgMeanValue(community,species)
getOrgSubstrate(community,species,elements,substrate,inter)
getTemp(community,element,sub)
getTime(community,species,x)
host.ecto(community,species,host,dead)
host.endo(community,species,host,harm)
init.timing(community)
initCount(community,species,file,append,debugit)
initOffspring(community,species)
initOrgInfo(package)
initTemp(community,lo.hour,hi.hour,days)
leftist.birth(organism,neworg,free)
leftist.create(tree,key)
leftist.merge(tree,P,Q,R,key)
leftist.remove(tree,P)
leftist.tree(x,tree,key,empty)
leftist.update(tree,P)
logpretty(x,n,roundit)
my.eval(species,extension,element,checkdata)
mydata(dataname,package,restart)
\method{plot}{current}(x,species,col,headstuff,units,right,adj,position,pch,cex,stage,xlab,ylab,future,...)
put.base(community,species,id,free)
ramp.backSpline(s)
readCount(community,species)
rescale.temp(low,high,newry,newrx,ry)
ring.add(ring,x)
ring.remove(ring,P)
rspline(meantime,fivepar,fit,meanvalue,invmvalue,span)
rtri(n,width,tri,roundoff)
sampleOrgSubstrate(community,species,elements,substrate.name,inter)
set.birth(community,species,neworg)
set.offspring(community,species,host,dead)
setCount(community,species,elements)
setEvents(community,period)
setOrgInfo(community,species,hosts,package)
setOrgMeanValue(community,species,stage,mvalue)
setTemp(community,element,value)
showTemp(community)
sierpinski(stage,reset)
spline.deriv(s)
spline.design(y,x,nspline,xy,n,horizontal)
spline.extrapolate(meanvalue,invmvalue,x)
spline.meanvalue(x,y,data,nspline,xy,tol,n)
spline.rate(meanvalue,x)
splinesum(xy,fit,log2,tol)
substrate.plot(community,count,mfcol,main,col,cex,logaxis,debugit)
summaryshow(xy,fit,col,sums)
temp.lines(s,mult,col)
temp.repeat(community,period)
temp.spline(community,hour,temp,start,mintemp,cumulative,mult)
text.current(x,species,col,pch,cex,position,stage,future,...)
transTime(community,org1name,org2name,x,unit1,unit2)
tri2car(tri,xmult,ymult,sq3)
tri2car.default(aa,bb,cc)
tridist(tri)
update.mintime(object,species,...)
updateCount(community,species,individual,is.death,step)
updateCounts(community,species,newbirths)
updateEvents(community,species,event,increment)
updateTemp(community,period)
writeCount(community,species,time,future,countage,countsub)
}
\author{ Brian S. Yandell, yandell@stat.wisc.edu }
\keyword{internal}
