% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing_ageclass.R
\name{ewing_ageclass}
\alias{ewing_ageclass}
\alias{ggplot_ewing_ageclass}
\alias{autoplot.ewing_ageclass}
\title{Ravel count from Ewing simulation models}
\usage{
ewing_ageclass(
  community,
  substrate = TRUE,
  total = TRUE,
  normalize = TRUE,
  ...
)

ggplot_ewing_ageclass(object, main = NULL, title = NULL, ...)

\method{autoplot}{ewing_ageclass}(object, ...)
}
\arguments{
\item{community}{object of class \code{ewing} with population data by
species}

\item{substrate}{include substrate if \code{TRUE}}

\item{total}{add total column if \code{TRUE}}

\item{normalize}{normalize counts to 1 if \code{TRUE}}

\item{...}{extra parameters not used}

\item{object}{object of class \code{ewing_ageclass}}

\item{main}{plot main title}

\item{title}{plot title}
}
\description{
Ravel counts to produce one comprehensive data frame across species
}
\examples{


\dontrun{
ewing_ageclass( community )
}


}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}},
\code{\link{summary.ewing}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
