% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing_discrete.R
\name{ewing_discrete1}
\alias{ewing_discrete1}
\alias{ewing_discrete}
\alias{make_ewing_discrete}
\alias{summary.ewing_discrete}
\title{Discrete time simulation for Ewing Envelope}
\usage{
ewing_discrete1(increment = 0.5, ...)
}
\arguments{
\item{increment}{increment for discrete simulation time}

\item{...}{additional parameters}

\item{nsim}{number of simulations to run}

\item{verbose}{show `.` for each simulation if `TRUE`}

\item{object}{object of class `ewing_envelope` or `ewing_envelopes`}
}
\description{
Do one simulation and save only by discrete `time` and `increment`
}
\details{
Wrapper to set up class and attributes of `ewing_envelope` object.
}
