% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{ewing_envelope}
\alias{ewing_envelope}
\title{Create Envelope of Ewing Simulations}
\usage{
ewing_envelope(object, species, item, ordinate = "time", increment = 0.5)
}
\arguments{
\item{object}{multiple simulation object from `envelope_sim`}

\item{species}{name of species in `object` to build envelope}

\item{item}{name of item in `species` to build envelope}

\item{ordinate}{name of ordinate (X axis) to build envelope}

\item{increment}{increament for discretizing}
}
\description{
Create envelope object for plotting from multiple runs of Ewing simulation.
}
