% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing_snapshot.R
\name{ewing_snapshot}
\alias{ewing_snapshot}
\title{Title}
\usage{
ewing_snapshot(object, step = 0, ...)
}
\arguments{
\item{object}{object of class `ewing`}

\item{step}{step number for snapshot}

\item{...}{additional arguments}
}
\value{
object of class ewing_snapshot
}
\description{
Title
}
