% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewingFutureApp.R
\name{ewingFutureApp}
\alias{ewingFutureApp}
\alias{ewingFutureServer}
\alias{ewingFutureInput}
\alias{ewingFutureOutput}
\title{Ewing App}
\usage{
ewingFutureApp(title = "Population Ethology")

ewingFutureServer(id, siminit)

ewingFutureInput(id)

ewingFutureOutput(id)
}
\description{
Ewing App

Ewing Input

Ewing Output
}
