% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewingInitApp.R
\name{ewingInitApp}
\alias{ewingInitApp}
\alias{ewingInitServer}
\alias{ewingInitInput}
\alias{ewingInitOutput}
\title{Ewing Initialization App}
\usage{
ewingInitApp(title = "Population Ethology")

ewingInitServer(id)

ewingInitInput(id)

ewingInitOutput(id)
}
\description{
Ewing Initialization App

Ewing Input

Ewing Output
}
