% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing.R
\name{ewingInput}
\alias{ewingInput}
\alias{ewingOutput}
\alias{ewingApp}
\title{Ewing Input}
\usage{
ewingInput(id)

ewingOutput(id)

ewingApp(title = "Population Ethology")
}
\description{
Ewing Input

Ewing Output

Ewing App
}
