% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivePlotApp.R
\name{fivePlotApp}
\alias{fivePlotApp}
\title{Interactive 5-Parameter Curve Explorer App}
\usage{
fivePlotApp(title = "Spline 5-Parameter Plot Explorer")
}
\arguments{
\item{title}{Application title}
}
\description{
Shiny application to interactively explore and visualize the 5-parameter curve adjustments using \code{five.plot}.
}
