% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveShowApp.R
\name{fiveShowApp}
\alias{fiveShowApp}
\title{Interactive Spline Goal Explorer App}
\usage{
fiveShowApp(title = "Spline 5-Parameter Goal Explorer")
}
\arguments{
\item{title}{Application title}
}
\description{
Shiny application to interactively explore and search for parameter adjustments using \code{five.show}.
}
