% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivetimeApp.R
\name{fivetimeApp}
\alias{fivetimeApp}
\title{Interactive Spline and Five-Parameter Explorer App}
\usage{
fivetimeApp(title = "Spline Time Parameter Explorer")
}
\arguments{
\item{title}{Application title}
}
\description{
Shiny application to interactively explore and visualize the spline-based time-temperature 
relationship and the 5-parameter curve adjustments using \code{five.plot} and \code{five.show}.
}
