% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureApp.R
\name{futureApp}
\alias{futureApp}
\alias{futureServer}
\alias{futureInput}
\alias{futureOutput}
\title{Ewing Future Events App}
\usage{
futureApp(title = "Population Ethology")

futureServer(id, siminit)

futureInput(id)

futureOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{siminit}{reactive state of an initialized `community` topological root}
}
\description{
Module tracking direct continuous horizon mappings specifically built on processing
deterministic `future.events()` sequences rather than discrete simulation iterations.
}
