% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{get_huc_from_point}
\alias{get_huc_from_point}
\title{Get HUC12 From Map Click Coordinate}
\usage{
get_huc_from_point(lng, lat)
}
\arguments{
\item{lng}{Numeric longitude coordinate}

\item{lat}{Numeric latitude coordinate}
}
\value{
An `sf` polygon representation of the covering HUC12.
}
\description{
Automatically retrieves the underlying USGS HUC12 subwatershed boundary polygon
given an arbitrary Longitude and Latitude pair (usually derived from a user click).
This powers the Option B interaction paradigm.
}
