% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_ewing}
\alias{ggplot_ewing}
\alias{autoplot.ewing}
\alias{ggplot_ewing_ageclass}
\alias{autoplot.ewing_ageclass}
\title{ggplot of Ewing simulation model age classes}
\usage{
ggplot_ewing(object, step = 0, ageclass = TRUE, substrate = TRUE, ...)

\method{autoplot}{ewing}(object, ...)

ggplot_ewing_ageclass(object, ...)

\method{autoplot}{ewing_ageclass}(object, ...)
}
\arguments{
\item{...}{other plot parameters}

\item{x}{object of class \code{ewing} with population data by species}
}
\description{
GGPlot of various aspects of simulation.
}
\examples{

\dontrun{
ggplot_ewing( community )
}

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}},
\code{\link{summary.ewing}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
