% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{ggplot_ewing_envelope}
\alias{ggplot_ewing_envelope}
\alias{autoplot.ewing_envelope}
\title{GGplot of Ewing envelope}
\usage{
ggplot_ewing_envelope(
  crset,
  cols = c("#21908CFF", "#440154FF", "#5DC863FF"),
  ...
)

\method{autoplot}{ewing_envelope}(object, ...)
}
\arguments{
\item{crset}{object of class `ewing_envelope`}

\item{cols}{colors for top simulations}

\item{...}{additional arguments}

\item{object}{object of class `ewing_envelope`}
}
\description{
GGplot of Ewing envelope

GGplot of Ewing Envelope
}
