% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing_envelope.R
\name{ggplot_ewing_envelopes}
\alias{ggplot_ewing_envelopes}
\alias{ggplot_ewing_envelope}
\alias{autoplot.ewing_envelope}
\title{GGplot of Ewing multiple envelopes}
\usage{
ggplot_ewing_envelopes(object, confidence = FALSE, main = "", ...)
}
\arguments{
\item{object}{object of class `ewing_envelope`}

\item{confidence}{plot confidence bands if `TRUE`}

\item{main}{title for plot}

\item{...}{additional arguments}

\item{cols}{colors for top simulations}
}
\description{
GGplot of Ewing multiple envelopes
}
\details{
GGplot of Ewing envelope

GGplot of Ewing Envelope
}
