% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initApp.R
\name{initApp}
\alias{initApp}
\alias{initServer}
\alias{initOutput}
\title{Ewing Initialization App}
\usage{
initApp(title = "Population Ethology")

initServer(id, init_par)

initOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{init_par}{list of natively structured initialization counts (e.g. `$host`, `$parasite`)}
}
\description{
A module rendering environment initialization frameworks independently out of standard runtime constraints.
Evaluates base configurations via `init.simulation()`.
}
