% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParApp.R
\name{initParApp}
\alias{initParApp}
\alias{initParServer}
\alias{initParInput}
\alias{initParUI}
\alias{initParOutput}
\title{Ewing Initialization App}
\usage{
initParApp(title = "Population Ethology")

initParServer(
  id,
  simres = shiny::reactiveVal(NULL),
  datafile = shiny::reactiveVal("")
)

initParInput(id)

initParUI(id)

initParOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{simres}{reactive simulation state used to structure dynamic data grid evaluations}

\item{datafile}{reactive filepath pointing to optional target parameter overwrites}
}
\description{
Shiny module capturing baseline evaluation traits and populating parameter inputs dynamically
based heavily on external matrix definitions.
}
