% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multApp.R
\name{multApp}
\alias{multApp}
\alias{multServer}
\alias{multInput}
\alias{multOutput}
\title{Multiple Simulation Envelope App}
\usage{
multApp(title = "Population Ethology")

multServer(id)

multInput(id)

multOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}
}
\description{
A specialized UI pipeline scaling the core population ethology topologies directly into 
extensive sequence horizons producing bounds and dynamic envelope metrics.
}
