% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origEwingApp.R
\name{origEwingApp}
\alias{origEwingApp}
\alias{origEwingServer}
\alias{origEwingInput}
\alias{origEwingOutput}
\title{Ewing App}
\usage{
origEwingApp(title = "Population Ethology")

origEwingServer(id)

origEwingInput(id)

origEwingOutput(id)
}
\description{
Ewing App

Ewing Input

Ewing Output
}
