% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ewing}
\alias{plot.ewing}
\title{plot of Ewing simulation models}
\usage{
\method{plot}{ewing}(x, substrate = TRUE, mfcol = mfcols, ..., ageclass = TRUE)
}
\arguments{
\item{x}{object of class \code{ewing} with population data by species}

\item{substrate, ageclass}{include `substrate` or `ageclass` plot if TRUE}

\item{mfcol}{par parameter reset by default}

\item{...}{other plot parameters}
}
\description{
Plot of various aspects of simulation.
}
\examples{

\dontrun{
plot.ewing( community )
}

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}},
\code{\link{summary.ewing}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
