% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ravel.count}
\alias{ravel.count}
\title{Ravel count from Ewing simulation models}
\usage{
ravel.count(community, substrate = TRUE, total = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{community}{object of class \code{ewing} with population data by species}

\item{substrate}{include substrate if \code{TRUE}}

\item{total}{add total column if \code{TRUE}}

\item{normalize}{normalize counts to 1 if \code{TRUE}}

\item{...}{extra parameters not used}
}
\description{
Ravel counts to produce one comprehensive data frame across species
}
\examples{

\dontrun{
ravel.count( community )
}

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}},
\code{\link{summary.ewing}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
