% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redscale.R
\name{redscale}
\alias{redscale}
\title{Red Scale simulation initiation}
\format{
\code{redscale} is a data frame with the following columns
\describe{
 \item{DD}{mean degree days to future event}
 \item{aphytis.feed}{relative risk of being killed by an Aphytis that is feeding}
 \item{aphytis}{relative risk of being killed by an Aphytis that is laying eggs}
 \item{comperiella}{relative risk for comperiella}
 \item{encarsia}{relative risk for encarsia}
 \item{gender}{gender (only females modeled here)}
 \item{pch}{plot character}
 \item{color}{plot color}
 \item{stage}{life stage of organism}
}
}
\description{
Sample simulation initiation setup.
}
\examples{

data(redscale)

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.population}}
}
\keyword{datasets}
