% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simApp.R
\name{simApp}
\alias{simApp}
\alias{simServer}
\alias{simInput}
\alias{simUI}
\title{Ewing Simulation Engine App}
\usage{
simApp(title = "Population Ethology Simulation")

simServer(id, init_par, datafile = shiny::reactiveVal(""))

simInput(id)

simUI(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{init_par}{reactive list mapping initial population sizes}

\item{datafile}{reactive matrix filepath containing physical environment and configuration targets}
}
\description{
Central evaluation module directing state flow interactions spanning across discrete evaluation limits
($nsim > 1$) and linear continuous event evaluations ($nsim == 1$).
}
