% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Simulated Data with Multiple Runs}
\format{
Object of class `ewing_discrete`.
Contains `nsim` runs each of `nstep` steps.
Attributes include the following
\describe{
 \item{nsim}{number of simulations}
 \item{nstep}{number of steps per simulation}
 \item{count}{vector with initial population count by species}
 \item{species}{bector of species names}
 \item{items}{list by species of age classes (used in summary and plot methods)}
 \item{ordinate}{ordinate for simulation sequesnce (usually `time`)}
}
}
\description{
Result of `ewing_discrete` with 100 runs of
}
\examples{

data(simdata)

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{ewing_discrete}}
}
\keyword{datasets}
