% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simpleInput}
\alias{simpleInput}
\alias{simpleOutput}
\alias{simpleApp}
\title{simple Input}
\usage{
simpleInput(id)

simpleOutput(id)

simpleApp(title = "Population Ethology")
}
\description{
simple Input

simple Output

simple App
}
