% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ewing.R
\name{summary.ewing}
\alias{summary.ewing}
\title{summary of Ewing simulation models}
\usage{
\method{summary}{ewing}(object, ...)
}
\arguments{
\item{object}{object of class \code{ewing} with population data by species}

\item{...}{other summary parameters}
}
\description{
Summary of various aspects of simulation.
}
\examples{


\dontrun{
summary.ewing( community )
}


}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}},
\code{\link{plot.ewing}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
