% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysetholApp.R
\name{sysetholApp}
\alias{sysetholApp}
\alias{sysetholInput}
\alias{sysetholOutput}
\alias{sysetholServer}
\title{Systems Ethology App}
\usage{
sysetholApp(title = "Systems Ethology Platform")

sysetholInput(id)

sysetholOutput(id)

sysetholServer(id)
}
\arguments{
\item{title}{Application title}

\item{id}{Module ID string}
}
\description{
Interactive Systems Ethology platform orchestrating host-parasite individual-based simulations,
age class distributions, hexagonal spatial substrate networks, variance envelopes, and input data tables.
}
