% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempApp.R
\name{tempApp}
\alias{tempApp}
\title{Interactive Daily Temperature Design Explorer App}
\usage{
tempApp(community = NULL, title = "Daily Temperature Design Explorer")
}
\arguments{
\item{community}{Simulation object of class \code{ewing} (default: \code{init.simulation()})}

\item{title}{Application title}
}
\description{
Shiny application to interactively adjust daily high and low temperature splines
and visualize degree-day accumulation dynamics for an \code{ewing} simulation.
}
