% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangleApp.R
\name{triangleApp}
\alias{triangleApp}
\title{Interactive Tridiagonal Substrate Network Explorer App}
\usage{
triangleApp(
  width = 10,
  step = 1,
  title = "Tridiagonal Substrate Network Explorer"
)
}
\arguments{
\item{width}{Integer radius size limit of spatial components (default: 10)}

\item{step}{Numeric grid density spacing interval (default: 1)}

\item{title}{Application title}
}
\description{
Shiny application to interactively build, visualize, and inspect tridiagonal
geometric substrate spatial networks (\code{substrate_topology}, \code{create_substrate}, \code{autoplot.substrate}).
}
