% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distPlotApp.R
\name{distPlotApp}
\alias{distPlotApp}
\alias{distPlotServer}
\alias{distPlotInput}
\alias{distPlotOutput}
\title{Distribution Plot App}
\usage{
distPlotApp(title = "Population Ethology")

distPlotServer(id, simres)

distPlotInput(id)

distPlotOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}

\item{simres}{reactive simulation state holding the active `ewing` data object}
}
\description{
Shiny UI components and module server logic for visualizing population age-class 
distribution distributions and demographic structures.
}
