% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distPlotApp.R
\name{distPlotApp}
\alias{distPlotApp}
\alias{distPlotServer}
\alias{distPlotInput}
\alias{distPlotOutput}
\title{Distribution Plot App}
\usage{
distPlotApp(title = "Population Ethology")

distPlotServer(id, simres)

distPlotInput(id)

distPlotOutput(id)
}
\description{
Distribution Plot App
}
