% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{ewing_discrete}
\alias{ewing_discrete}
\title{Envelope for Ewing simulations}
\usage{
ewing_discrete(nsim, increment = 1)
}
\arguments{
\item{nsim}{number of simulations to run}

\item{increment}{increment for discrete simulation time}
}
\description{
This currently relies on integer `step` value, but ultimately want to use the `time`,
which are not the same across simulations. That will require using spline interpolation
to a common grid. More work.
 
https://cran.r-project.org/web/packages/GET/GET.pdf
https://cran.r-project.org/web/packages/GET/vignettes/GET.pdf
https://medium.com/runic-software/simple-guide-to-plotting-monte-carlo-envelopes-for-spatial-statistics-in-r-295b04b64ec1
https://rss.onlinelibrary.wiley.com/doi/10.1111/j.1467-9876.2009.00701.x
https://stat.ethz.ch/R-manual/R-devel/library/boot/html/envelope.html
}
