% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing_discrete.R, R/make_ewing_discrete.R
\name{ewing_discrete}
\alias{ewing_discrete}
\alias{summary.ewing_discrete}
\alias{make_ewing_discrete}
\title{Envelope for Ewing simulations}
\usage{
ewing_discrete(nsim, verbose = FALSE, ...)

\method{summary}{ewing_discrete}(object, ...)

make_ewing_discrete(object)
}
\arguments{
\item{nsim}{number of simulations to run}

\item{verbose}{show `.` for each simulation if `TRUE`}

\item{...}{additional parameters}

\item{object}{list of objects from `ewing_discrete`}
}
\description{
Wrapper to set up class and attributes of `ewing_envelope` object.
}
