% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{ewing_discrete1}
\alias{ewing_discrete1}
\alias{ewing_discrete}
\alias{make_ewing_discrete}
\alias{summary.ewing_discrete}
\title{Discrete time simulation for Ewing Envelope}
\usage{
ewing_discrete1(increment = 0.5, ...)

ewing_discrete(nsim, ...)

make_ewing_discrete(object)

\method{summary}{ewing_discrete}(object, ...)
}
\arguments{
\item{increment}{increment for discrete simulation time}

\item{...}{additional parameters}

\item{nsim}{number of simulations to run}

\item{object}{object of class `ewing_envelope` or `ewing_envelopes`}
}
\description{
Do one simulation and save only by discrete `time` and `increment`

Wrapper to set up class and attributes of `ewing_envelope` object.
}
