% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewing_envelope.R
\name{ewing_envelope}
\alias{ewing_envelope}
\alias{ewing_envelopes}
\alias{summary.ewing_envelopes}
\alias{print.ewing_envelopes}
\title{Create Envelope of Ewing Simulations}
\usage{
ewing_envelope(object, species, item, ordinate = "time", increment = 0.5)

ewing_envelopes(object)

\method{summary}{ewing_envelopes}(object, species = NULL, verbose = TRUE, ...)

\method{print}{ewing_envelopes}(x, species = NULL, ...)
}
\arguments{
\item{object}{object of class `ewing_envelope` or `ewing_envelopes`}

\item{species}{subset on `species` if not `NULL`}

\item{item}{name of item in `species` to build envelope}

\item{ordinate}{name of ordinate (X axis) to build envelope}

\item{increment}{increament for discretizing}

\item{verbose}{print settings if `TRUE`}

\item{...}{additional parameters}

\item{x}{object of class `ewing_envelope` or `ewing_envelopes`}
}
\description{
Create envelope object for plotting from multiple runs of Ewing simulation.
}
