% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{ewing_substrate}
\alias{ewing_substrate}
\alias{ggplot_ewing_substrate}
\alias{autoplot.ewing_substrate}
\title{Ewing Substrate by Species}
\usage{
ewing_substrate(
  community,
  species,
  headstuff = c(0, "start", sum(to.plot)),
  units = getOrgFeature(community, species, "units"),
  right = species,
  adj = c(0, 0.5, 1),
  show_sub = substrates,
  step = 0,
  ...
)

ggplot_ewing_substrate(object, xlab = "horizontal", ylab = "vertical", ...)

\method{autoplot}{ewing_substrate}(object, ...)
}
\description{
Ewing Substrate by Species
}
