% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewingApp.R
\name{ewingApp}
\alias{ewingApp}
\alias{ewingServer}
\alias{ewingInput}
\alias{ewingOutput}
\title{Ewing App}
\usage{
ewingApp(title = "Systems Ethology")

ewingServer(id)

ewingInput(id)

ewingOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}
}
\description{
Core `Systems Ethology` Shiny platform acting as the central interactive wrapper modularizing
age distributions, probability envelopes, and runtime data interactions.
}
