% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewingApp.R
\name{ewingApp}
\alias{ewingApp}
\alias{ewingServer}
\alias{ewingInput}
\alias{ewingOutput}
\title{Ewing App}
\usage{
ewingApp(title = "Population Ethology")

ewingServer(id)

ewingInput(id)

ewingOutput(id)
}
\description{
Ewing App

Ewing Input

Ewing Output
}
