% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.meanvalue.R
\name{future.meanvalue}
\alias{future.meanvalue}
\title{design mean value curves for future events in Ewing simulation models}
\usage{
future.meanvalue(community, species, event = future$current[1], data)
}
\arguments{
\item{community}{object with population data by species}

\item{species}{name of species}

\item{event}{name of future event}

\item{data}{data if available}
}
\description{
Uses interactive spline tool to design curves.
}
\examples{


\dontrun{
future.meanvalue( community, "host", "first.instar" )
}


}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
