% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureApp.R
\name{futureApp}
\alias{futureApp}
\alias{futureServer}
\alias{futureInput}
\alias{futureOutput}
\title{Ewing App}
\usage{
futureApp(title = "Population Ethology")

futureServer(id, siminit)

futureInput(id)

futureOutput(id)
}
\description{
Ewing App

Ewing Input

Ewing Output
}
