% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_ewing}
\alias{ggplot_ewing}
\title{ggplot of Ewing simulation models}
\usage{
ggplot_ewing(x, ...)
}
\arguments{
\item{x}{object of class \code{ewing} with population data by species}

\item{...}{other plot parameters}
}
\description{
GGPlot of various aspects of simulation.
}
\examples{

\dontrun{
ggplot_ewing( community )
}

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.simulation}}, \code{\link{future.events}},
\code{\link{summary.ewing}}
}
\author{
Brian S. Yandell, \email{yandell@stat.wisc.edu}
}
\keyword{utilities}
