% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{autoplot.ewing_envelope}
\alias{autoplot.ewing_envelope}
\title{GGplot of Ewing Envelope}
\usage{
\method{autoplot}{ewing_envelope}(object, ...)
}
\arguments{
\item{object}{object of class `ewing_envelope`}

\item{...}{additional arguments}
}
\description{
GGplot of Ewing Envelope
}
