% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.population}
\alias{init.population}
\title{Initialize Organism Population}
\usage{
init.population(
  community,
  species,
  n = 200,
  width = 100,
  units = getOrgFeature(community, species, "units"),
  timeit = FALSE,
  reject = Inf,
  position = rtri(n, width),
  colnames = c(leftistnames, paramnames, posnames, eventnames),
  init.stage = istage,
  init.weight = getOrgFuture(community, species, "init"),
  messages = TRUE,
  ...
)
}
\arguments{
\item{community}{community structure}

\item{species}{species name(s) as character string vector}

\item{n}{number of organisms in population}

\item{width}{maximum width of dispersal on tridiagonal coordinate system}

\item{units}{time units (default is
\code{getOrgFeature(community,species,"units")})}

\item{timeit}{print timing information if \code{TRUE}}

\item{reject}{rejection time (default is \code{Inf})}

\item{position}{initial positions in triangular coordinate system (generated
randomly)}

\item{colnames}{names of columns in data structure (do not change)}

\item{init.stage}{initial stage of all organisms (default values generated
randomly)}

\item{init.weight}{initial weights for life stages (default taken for
organism features)}

\item{...}{not used}
}
\value{
\item{comp1 }{Description of `comp1'} \item{comp2 }{Description of
`comp2'} ...
}
\description{
Create data structure with n organisms distributed across life stages.
Includes leftist tree structure
}
\details{
Does various initializations based in information in \code{future.xxx},
where \code{xxx} is replaced by the name of the dataset. Normally,
\code{init.population} is called by \code{init.simulation}.
}
\examples{

\dontrun{init.population( redscale )}

}
\seealso{
\code{\link{init.simulation}}
}
\author{
Brian S. Yandell
}
\keyword{utilities}
