% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.simulation}
\alias{init.simulation}
\title{Initialize simulation run}
\usage{
init.simulation(
  package = "ewing",
  count = 200,
  interact = FALSE,
  messages = TRUE,
  ...
)
}
\arguments{
\item{package}{package where community data features can be found}

\item{count}{number of individuals per species (single number or count in order of `species`)}

\item{interact}{ask for species count if `TRUE` and interactive}

\item{messages}{show messages if `TRUE` (default)}

\item{...}{additional arguments for `init.population`}
}
\value{
Object with elements for each species that are created by init.population.
}
\description{
Initializes simulation using organism features.
Default data are used unless user provides global data
or uses the hidden argument `datadir` to specify a folder with user data.
}
\examples{

\dontrun{init.simulation(myrun)}

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.population}}
}
\author{
Brian S. Yandell
}
\keyword{utilities}
