% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.simulation}
\alias{init.simulation}
\title{Initialize simulation run}
\usage{
init.simulation(
  package = "ewing",
  species = getOrgFeature(community)[1:2],
  hosts = getOrgHosts(community, species),
  count = 200,
  interact = interactive()
)
}
\arguments{
\item{package}{package where community data features can be found}

\item{species}{list of species to simulate}

\item{hosts}{list of hosts for species}

\item{count}{number of individuals per species}

\item{interact}{ask for species count if true and interactive}
}
\value{
Object with elements for each species that are created by
init.population.
}
\description{
Initializes simulation using organizm features.
}
\examples{

\dontrun{init.simulation(myrun)}

}
\references{
\url{www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link{init.population}}
}
\author{
Brian S. Yandell
}
\keyword{utilities}
