% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initApp.R
\name{initApp}
\alias{initApp}
\alias{initServer}
\alias{initOutput}
\title{Ewing Initialization App}
\usage{
initApp(title = "Population Ethology")

initServer(id, init_par)

initOutput(id)
}
\description{
Ewing Initialization App
}
