% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParApp.R
\name{initParApp}
\alias{initParApp}
\alias{initParServer}
\alias{initParInput}
\alias{initParUI}
\alias{initParOutput}
\title{Ewing Initialization App}
\usage{
initParApp(title = "Population Ethology")

initParServer(id)

initParInput(id)

initParUI(id)

initParOutput(id)
}
\description{
Ewing Initialization App
}
