% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletApp.R
\name{leafletOutput}
\alias{leafletOutput}
\alias{leafletServer}
\alias{leafletApp}
\title{Interactive Leaflet Mapping UI (Output)}
\usage{
leafletOutput(id)

leafletServer(id)

leafletApp()
}
\arguments{
\item{id}{Module ID}
}
\value{
A list of reactive objects: `huc` (reactiveVal holding discovered `sf` HUC polygon(s)),
  `status` (reactiveVal holding HTML status message), `click` (reactive holding map click details),
  and `drawn_polygon` (reactiveVal holding user drawn rubberband polygon sf).
}
\description{
Server logic for interactive Leaflet discovery. Returns a list of reactives
(`huc`, `status`, `click`, `drawn_polygon`) enabling Shiny module composition.
}
