% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletApp.R
\name{leafletInput}
\alias{leafletInput}
\alias{leafletOutput}
\alias{leafletServer}
\alias{leafletApp}
\title{Interactive Leaflet Mapping UI (Input)}
\usage{
leafletInput(id)

leafletOutput(id)

leafletServer(id)

leafletApp()
}
\arguments{
\item{id}{Module ID}
}
\description{
Interactive Leaflet Mapping UI (Input)

Interactive Leaflet Mapping UI (Output)

Interactive Leaflet Mapping Server Logic

Run the Leaflet Mapping Integration App Tracker
}
