% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multApp.R
\name{multApp}
\alias{multApp}
\alias{multServer}
\alias{multInput}
\alias{multOutput}
\title{Ewing App}
\usage{
multApp(title = "Population Ethology")

multServer(id)

multInput(id)

multOutput(id)
}
\description{
Ewing App

Ewing Input

Ewing Output
}
