% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrateApp.R
\name{substrateApp}
\alias{substrateApp}
\alias{substrateServer}
\alias{substrateOutput}
\title{Substrate Plot App}
\usage{
substrateApp()

substrateServer(id, simres)

substrateOutput(id)
}
\arguments{
\item{id}{module ID string}

\item{simres}{reactive object resolving chronological population metrics}

\item{title}{Application title}
}
\description{
A focused graphical module aggregating structural species progression separated distinctly across
physical environment thresholds mapped by evaluating top-level topological definitions.
}
