% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrateApp.R
\name{substrateApp}
\alias{substrateApp}
\alias{substrateServer}
\alias{substrateOutput}
\title{Substrate Plot App}
\usage{
substrateApp()

substrateServer(id, simres)

substrateOutput(id)
}
\description{
Substrate Plot App
}
