% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.design.R
\name{temp.design}
\alias{temp.design}
\alias{temp.plot}
\title{graphical adjustment of daily high and low temperature}
\usage{
temp.design(
  community,
  nspline = 8,
  n = 1,
  horizontal = TRUE,
  col = c(low = "blue", high = "red")
)
}
\arguments{
\item{community}{object with population data by species}

\item{nspline}{number of spline nodes}

\item{n}{number of steps to perform}

\item{horizontal}{use horizontal orientation if TRUE}

\item{col}{colors of \code{low} and \code{high} temperature ranges}

\item{lo.hour}{lower limit of hours to plot}

\item{hi.hour}{upper limit of hours to plot}

\item{length}{number of interpolation points for plot}

\item{derivative}{show derivative of temperature}

\item{...}{other arguments to pass to plot}

\item{printit}{print out hours and temperatures if \code{TRUE}}
}
\description{
Graphical interface updates daily high and low temperatures, along with
rescaling the days and temperature range.
}
\examples{


\dontrun{
temp.design()
temp.plot()
}


}
\references{
See \url{www.stat.wisc.edu/~yandell/ewing}.
}
\seealso{
\code{\link{TemperatureBase}}
}
\author{
Brian S. Yandell
}
\keyword{utilities}
