% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershedApp.R
\name{watershedApp}
\alias{watershedApp}
\alias{watershedServer}
\alias{watershedInput}
\alias{watershedOutput}
\title{Watershed App}
\usage{
watershedApp(title = "Ewing Watershed Projection")

watershedServer(id)

watershedInput(id)

watershedOutput(id)
}
\arguments{
\item{title}{Application title}

\item{id}{module ID string}
}
\description{
Real-time USGS topological projection interfaces capable of drawing geometric features statically
onto active coordinate matrices dynamically.
}
